Check engine requirements for Gradle support
This plugin is based on Cordova, which is used for Meteor's mobile apps,
Chrome Mobile Apps, Phonegap, Steroids, and others. Not all of these
tools are updated to use Cordova 5 and cordova-android 4, which allows
Gradle build steps to execute. As such, dependencies are not downloaded
and the plugin fails to execute.

An alternative would be to bundle the dependencies like iOS. This has
its own set of problems, but would ensure broader support.

Close #6